[META 388] Proposal for collecting Azure App Service cloud metadata (#365)
This commit updates the cloud metadata
spec to propose collecting Azure App Service
metadata.

Azure App Services are a very popular PaaS offering
on Azure. In contrast to Azure VMs, App Services do
not have access to the internal metadata endpoint
in order to retrieve metadata about the app instance.
Instead, much of the metadata that is relevant to the
purposes of APM are available in environment variables.

The environment variables of interest are documented
here.