Add note about process transaction
This commit adds details about a process transaction to the
messaging spec. The message processing flow may be
distinct from the reception of message(s), so add guidance
for naming of a "process" transaction.

The scenario that brought about this PR
arises in instrumenting the Azure Service Bus
SDK for .NET. The reception of messsages is
a distinct event to message processing, that
starts and ends before the message processing
event occurs. In this instrumentation, the
reception of messages would be a RECEIVE
transaction followed by a PROCESS transaction.