Change default APM server_url to 'http://127.0.0.1:8200' to avoid IPv6 ambiguity
The current default APM server URL -- 'http://localhost:8200' -- is
ambiguous. "localhost" can resolve to a '127.0.0.1' (IPv4) or '[::1]'
(IPv6).

At least in Node v17 the default `dns.lookup()` ordering of results was
changed to no longer explicitly sort IPv4 addresses first.  That means
that on systems where the default resolver returns '[::1]' first for
"localhost" will result in a broken attempt to talk to a
default-configured local APM server:
  https://github.com/elastic/apm-agent-nodejs/issues/3045
This is because APM server only binds to the IPv4 port:
  https://github.com/elastic/apm-server/issues/1405